+++ eras/dijkstra/impl/testlib/Test/Cardano/Ledger/Dijkstra/Examples.hs
@@ -21,6 +21,14 @@ module Test.Cardano.Ledger.Dijkstra.Examples (
   exampleDijkstraGenesis,
 ) where
 
+import Cardano.Crypto.DSIGN (
+  BLS12381MinSigDSIGN,
+  DSIGNAggregatable (createPossessionProofDSIGN),
+  DSIGNAlgorithm (deriveVerKeyDSIGN, genKeyDSIGNWithContext),
+  seedSizeDSIGN,
+ )
+import Cardano.Crypto.DSIGN.BLS12381.Internal (minSigPoPDST)
+import Cardano.Crypto.Seed (mkSeedFromBytes)
 import Cardano.Ledger.Address (DirectDeposits (..))
 import Cardano.Ledger.Alonzo.Plutus.Context (EraPlutusTxInfo)
 import Cardano.Ledger.Alonzo.Scripts (ExUnits (..))
@@ -73,11 +81,19 @@ import Cardano.Ledger.Plutus.Data (
   dataToBinaryData,
  )
 import Cardano.Ledger.Plutus.Language (Language (..), plutusBinary)
+import Cardano.Ledger.State (
+  LeiosKey (..),
+  LeiosPossessionProof (..),
+  LeiosPubKey (..),
+  StakePoolParams (..),
+ )
+import qualified Data.ByteString as Strict
 import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as Map
 import Data.Maybe (fromJust)
 import qualified Data.OMap.Strict as OMap
 import qualified Data.OSet.Strict as OSet
+import Data.Proxy (Proxy (..))
 import qualified Data.Sequence.Strict as StrictSeq
 import Lens.Micro ((%~), (&), (.~), (<>~))
 import qualified PlutusLedgerApi.Common as P
@@ -102,6 +118,7 @@ import Test.Cardano.Ledger.Shelley.Examples (
   addShelleyBasedTopTxExampleFee,
   examplePayKey,
   exampleStakeKey,
+  exampleStakePoolParams,
   mkKeyHash,
   mkScriptHash,
  )
@@ -247,7 +264,15 @@ addDijkstraBasedTxFeatures tx =
         ]
     & bodyTxL . directDepositsTxBodyL .~ exampleDirectDeposits
     & bodyTxL . accountBalanceIntervalsTxBodyL .~ exampleAccountBalanceIntervals
+    & bodyTxL . certsTxBodyL
+      <>~ StrictSeq.fromList
+        [ RegPoolTxCert exampleStakePoolParamsWithLeiosKey
+        ]
   where
+    exampleStakePoolParamsWithLeiosKey =
+      exampleStakePoolParams
+        { sppLeiosKey = SJust exampleLeiosKey
+        }
     redeemers =
       Redeemers $
         Map.fromList
@@ -290,3 +315,14 @@ exampleDijkstraOnwardsEraPParamsUpdate =
     & ppuMaxRefScriptSizePerTxL .~ SJust (200 * 1024)
     & ppuRefScriptCostStrideL .~ SJust (knownNonZeroBounded @25_600)
     & ppuRefScriptCostMultiplierL .~ SJust (12 %! 10)
+
+exampleLeiosKey :: LeiosKey
+exampleLeiosKey =
+  LeiosKey
+    { leiosPubKey = LeiosPubKey vk
+    , leiosPossessionProof = LeiosPossessionProof $ createPossessionProofDSIGN minSigPoPDST sk
+    }
+  where
+    seed = mkSeedFromBytes $ Strict.replicate (fromIntegral $ seedSizeDSIGN (Proxy @BLS12381MinSigDSIGN)) 42
+    sk = genKeyDSIGNWithContext @BLS12381MinSigDSIGN Nothing seed
+    vk = deriveVerKeyDSIGN sk