+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -38,6 +38,13 @@ module Test.Cardano.Ledger.Core.Arbitrary (
 ) where
 
 import qualified Cardano.Chain.Common as Byron
+import Cardano.Crypto.DSIGN (
+  BLS12381MinSigDSIGN,
+  DSIGNAggregatable (PossessionProofDSIGN, createPossessionProofDSIGN),
+  DSIGNAlgorithm (genKeyDSIGNWithContext),
+  seedSizeDSIGN,
+ )
+import Cardano.Crypto.DSIGN.BLS12381.Internal (minSigPoPDST)
 import Cardano.Crypto.Hash.Class
 import Cardano.Ledger.Address
 import Cardano.Ledger.BaseTypes (
@@ -94,6 +101,7 @@ import Cardano.Ledger.TxIn (TxId (..), TxIn (..))
 import Control.Monad (replicateM)
 import Control.Monad.Identity (Identity)
 import Control.Monad.Trans.Fail.String (errorFail)
+import Data.Containers.ListUtils (nubOrdOn)
 import Data.Foldable (toList)
 import Data.GenValidity
 import Data.Int (Int64)
@@ -121,6 +129,7 @@ import Test.Cardano.Slotting.Arbitrary ()
 import Test.Cardano.StrictContainers.Instances ()
 import Test.Crypto.Hash ()
 import Test.Crypto.KES ()
+import Test.Crypto.Util (arbitrarySeedOfSize)
 import Test.Crypto.VRF ()
 import Test.QuickCheck
 import Test.QuickCheck.Arbitrary (GSubterms, RecursivelyShrink)
@@ -457,6 +466,7 @@ instance Arbitrary StakePoolParams where
     StakePoolParams
       <$> arbitrary
       <*> arbitrary
+      <*> pure BaseTypes.SNothing -- sppLeiosKey: only encoded in PV12+
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
@@ -478,9 +488,27 @@ instance Arbitrary StakePoolState where
       <*> arbitrary
       <*> arbitrary
       <*> arbitrary
+      <*> arbitrary
 
 instance Arbitrary PoolMetadata where
   arbitrary = PoolMetadata <$> arbitrary <*> arbitrary
+
+instance Arbitrary LeiosKey where
+  arbitrary = LeiosKey <$> arbitrary <*> arbitrary
+
+instance Arbitrary LeiosPubKey where
+  arbitrary = LeiosPubKey <$> arbitrary
+
+instance Arbitrary LeiosPossessionProof where
+  arbitrary = LeiosPossessionProof <$> genLeiosPossessionProof
+
+-- TODO: Replace with `defaultPossessionProofGen` once cardano-crypto-class
+-- exports it from the public API.
+genLeiosPossessionProof :: Gen (PossessionProofDSIGN BLS12381MinSigDSIGN)
+genLeiosPossessionProof = do
+  seed <- arbitrarySeedOfSize (seedSizeDSIGN (Proxy @BLS12381MinSigDSIGN))
+  let sk = genKeyDSIGNWithContext @BLS12381MinSigDSIGN Nothing seed
+  pure $ createPossessionProofDSIGN minSigPoPDST sk
 
 instance Arbitrary StakePoolRelay where
   arbitrary = genericArbitraryU
@@ -833,7 +861,8 @@ instance Arbitrary PV1.Data where
             oneof
               [ PV1.I <$> arbitrary
               , PV1.B <$> arbitrary
-              , PV1.Map <$> listOf ((,) <$> genData (n `div` 2) <*> genData (n `div` 2))
+              , PV1.Map . nubOrdOn fst
+                  <$> listOf ((,) <$> genData (n `div` 2) <*> genData (n `div` 2))
               , PV1.Constr
                   <$> fmap fromIntegral (arbitrary :: Gen Natural)
                   <*> listOf (genData (n `div` 2))